Add method to looking in every letter the most common array length
Add 'chars' to the font objects, this property has all characters
that that font can draw, this is a easier way to pass throug all
character arrays and look for the most common length. For now we
need to add every new character to this property, but when i develop
the font editor will be better.